feat: apply changed transform name planning
